0x08. C - Recursion TASKS

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -5,7 +5,7 @@
  * @m: Number.
  * Return: m.
  */
-int pow_number(int n, int m)
+int pow_number(long int n, long int m)
 {
 	if (m * m == n)
 		return (m);
